update: serviceを単一ファイルに変更
